# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Deception Chess API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface GameInput {
  playerName: string;
}

export interface JoinInput {
  playerName: string;
}

export interface ImpostorInput {
  playerId: string;
  /** The square of the enemy pawn selected as impostor (e.g. 'a7') */
  pawnSquare: string;
}

export interface ImpostorResult {
  success: boolean;
  bothReady: boolean;
}

export type GameJoinColor = typeof GameJoinColor[keyof typeof GameJoinColor];


export const GameJoinColor = {
  white: 'white',
  black: 'black',
} as const;

export interface GameJoin {
  gameId: string;
  playerId: string;
  color: GameJoinColor;
}

export type GameSummaryStatus = typeof GameSummaryStatus[keyof typeof GameSummaryStatus];


export const GameSummaryStatus = {
  waiting: 'waiting',
  selecting: 'selecting',
  active: 'active',
  finished: 'finished',
} as const;

export interface GameSummary {
  id: string;
  status: GameSummaryStatus;
  whitePlayerName: string;
  /** @nullable */
  blackPlayerName?: string | null;
  createdAt: string;
}

export interface PieceInfo {
  type: string;
  color: string;
  isImpostor: boolean;
  isSecured: boolean;
  impostorUsed: boolean;
}

export type GameStateStatus = typeof GameStateStatus[keyof typeof GameStateStatus];


export const GameStateStatus = {
  waiting: 'waiting',
  selecting: 'selecting',
  active: 'active',
  finished: 'finished',
} as const;

export type GameStateTurn = typeof GameStateTurn[keyof typeof GameStateTurn];


export const GameStateTurn = {
  white: 'white',
  black: 'black',
} as const;

/**
 * @nullable
 */
export type GameStateWinner = typeof GameStateWinner[keyof typeof GameStateWinner] | null;


export const GameStateWinner = {
  white: 'white',
  black: 'black',
  draw: 'draw',
} as const;

export interface GameState {
  id: string;
  status: GameStateStatus;
  /** FEN string of the current board position */
  fen: string;
  turn: GameStateTurn;
  whitePlayerName: string;
  /** @nullable */
  blackPlayerName?: string | null;
  /**
     * The requesting player's color (null if spectator)
     * @nullable
     */
  myColor?: string | null;
  /**
     * The square of the player's own impostor (only their own)
     * @nullable
     */
  myImpostorSquare?: string | null;
  /** Whether impostor selections are complete and game can start */
  impostorReady: boolean;
  whiteImpostorUsed: boolean;
  blackImpostorUsed: boolean;
  /**
     * Square of revealed white impostor (null if still hidden)
     * @nullable
     */
  whiteImpostorRevealed?: string | null;
  /**
     * Square of revealed black impostor (null if still hidden)
     * @nullable
     */
  blackImpostorRevealed?: string | null;
  whiteInvestigationUsed: boolean;
  blackInvestigationUsed: boolean;
  /** Squares of secured pawns */
  securedSquares: string[];
  /**
     * Description of last special event
     * @nullable
     */
  lastEvent?: string | null;
  /** @nullable */
  winner?: GameStateWinner;
  moveCount: number;
}