# lib/api-zod/src/generated/types/gameJoin.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Deception Chess API
 * OpenAPI spec version: 0.1.0
 */
import type { GameJoinColor } from './gameJoinColor';

export interface GameJoin {
  gameId: string;
  playerId: string;
  color: GameJoinColor;
}