# lib/api-zod/src/generated/types/gameState.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Deception Chess API
 * OpenAPI spec version: 0.1.0
 */
import type { GameStateStatus } from './gameStateStatus';
import type { GameStateTurn } from './gameStateTurn';
import type { GameStateWinner } from './gameStateWinner';

export interface GameState {
  id: string;
  status: GameStateStatus;
  /** FEN string of the current board position */
  fen: string;
  turn: GameStateTurn;
  whitePlayerName: string;
  /** @nullable */
  blackPlayerName?: string | null;
  /**
     * The requesting player's color (null if spectator)
     * @nullable
     */
  myColor?: string | null;
  /**
     * The square of the player's own impostor (only their own)
     * @nullable
     */
  myImpostorSquare?: string | null;
  /** Whether impostor selections are complete and game can start */
  impostorReady: boolean;
  whiteImpostorUsed: boolean;
  blackImpostorUsed: boolean;
  /**
     * Square of revealed white impostor (null if still hidden)
     * @nullable
     */
  whiteImpostorRevealed?: string | null;
  /**
     * Square of revealed black impostor (null if still hidden)
     * @nullable
     */
  blackImpostorRevealed?: string | null;
  whiteInvestigationUsed: boolean;
  blackInvestigationUsed: boolean;
  /** Squares of secured pawns */
  securedSquares: string[];
  /**
     * Description of last special event
     * @nullable
     */
  lastEvent?: string | null;
  /** @nullable */
  winner?: GameStateWinner;
  moveCount: number;
}