# lib/api-zod/src/generated/types/gameSummary.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Deception Chess API
 * OpenAPI spec version: 0.1.0
 */
import type { GameSummaryStatus } from './gameSummaryStatus';

export interface GameSummary {
  id: string;
  status: GameSummaryStatus;
  whitePlayerName: string;
  /** @nullable */
  blackPlayerName?: string | null;
  createdAt: string;
}