# lib/api-zod/src/generated/types/impostorInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Deception Chess API
 * OpenAPI spec version: 0.1.0
 */

export interface ImpostorInput {
  playerId: string;
  /** The square of the enemy pawn selected as impostor (e.g. 'a7') */
  pawnSquare: string;
}