# lib/api-zod/src/generated/types/impostorResult.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Deception Chess API
 * OpenAPI spec version: 0.1.0
 */

export interface ImpostorResult {
  success: boolean;
  bothReady: boolean;
}