# lib/api-zod/src/generated/types/pieceInfo.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Deception Chess API
 * OpenAPI spec version: 0.1.0
 */

export interface PieceInfo {
  type: string;
  color: string;
  isImpostor: boolean;
  isSecured: boolean;
  impostorUsed: boolean;
}